Save AC redirects before swapping AC uses
If the last use of an AC is swapped in this method,
the ref_count will be 0 and the AC row will be deleted.
If this happened, the code never got to add the AC redirects
to the appropriate table, since the ACs were completely gone by then.
This moves the AC redirect changes to happen *before*
the AC uses are changed.