Use index, not ID, to enable/disable annotation comparison
Apparently, a smaller ID does *not* guarantee an older annotation.
For an example of this, see release/115fd679-2135-4aea-af52-d5ba80574e05
This just uses the index of the annotation list for enabling and
disabling, since the annotations are sorted.
Keep in mind that the annotations are sorted from newest to oldest,
so the order of comparison is the opposite than it was with IDs.