MBS-12456: Load editor for latest_annotation on annotation page
We only create a sanitizedLatestAnnotation if we have editor data for it,
but we weren't loading that data on /annotation nor /annotation/id pages
so those never had a hydrated latest annotation to compare with,
and they didn't know when they were showing the latest annotation.
This ensures the editor is loaded, and should thus ensure
that the pages correctly indicate that this is the latest annotation
rather than an older revision.