MBS-12403 (II): Define and generate first diagrams
These diagrams are generated using the command `make` for the 2022 Q2
database schema, that is database schema sequence 27.

Some are drop-in replacement for old diagrams, others are just new.
Note that the diagrams are less compact than

- cdstub_details:

  Replacement for https://wiki.musicbrainz.org/File:cdstubs.png

- cover_art_details:

  Replacement for https://wiki.musicbrainz.org/File:cover_art.png

- entity_network_details:

  Replacement for https://wiki.musicbrainz.org/File:ngs.png

  It features new core entities: event, genre and series.
  It doesn’t feature alias table structure anymore, which is the same
  for each entity type, so as to focus on their specificities.
  General table strucure for core entity types is shown separately,
  as well as connections between core entity types, as follows:

  - entity_network_overview:

    It shows all core entity types and their connections but
    relationships through direct and indirect references if any.
    It focuses on core entity connections and hide the rest.

  - work_entity_details:

    It shows the companion tables for the core entity type “work”.
    It focuses on core entity properties and hide the rest.
    Similar diagrams can be added for each entity type.

- relationship_details:

  1st-half replacement for https://wiki.musicbrainz.org/File:ngs-ars.png

  It focuses on relationship’s own properties.

- relationship_overview:

  2nd-half replacement for https://wiki.musicbrainz.org/File:ngs-ars.png

  It focuses on relationship’s connections to core entities.

See "_comments" entries in the JSON file for further rationale behind
each diagram.

Note that no exhaustive diagram has NOT been added as it has shown to be
impossible to read due to the many lines overlap that cannot be followed
and to the very large number of tables making these lines very long too.

Also tables which are neither replicated nor available for download
(usually because it contains editor data) are obviously not included.