MBS-12470: Fix work language query for null
We were joining to work_language here, but obviously
that will only exist if there is at least one language set.
Also, there's no point having a FULL OUTER JOIN if we're only
returning rows that have a language in the WHERE.
The point of this was to also list works without a language set.