Fix warning when displaying blank annotation text
Annotation text is stored as `Maybe[Str]` in the model
`Entity/Annotation.pm` thus returning as undefined if blank.
This triggered the following warning when displaying annotation edits:

    Use of uninitialized value $old_annotation in concatenation (.) or
    string at lib/MusicBrainz/Server/Edit/Annotation/Edit.pm line 81.

This patch appropriately defines `$old_annotation` so that it can be
directly used as a string in the rest of the subroutine.