eslintrc: Set import/extensions to [error, always]
Extensions are required when importing ES modules in Node.js (and the
browser, where supported):
https://nodejs.org/api/esm.html#mandatory-file-extensions

This moves it to the unfixed rule list, since it's a reversal of the
previous rule and is definitely not fixed.  I changed warn to error
because a missing extension would actually cause an error in these
environments.