Remove clearRequireCache from root/server
In CommonJS land, Node.js caches modules that are require()'d into
require.cache.  If you clear a module from here, requiring it again will
reload it.

clearRequireCache is used in development to clear require.cache before
each request in the template renderer.  So if server-components.js gets
rebuilt, there's no need to restart the template renderer.

If we switch to ES modules in Node.js, there's no such thing as
require.cache, so this feature will no longer be possible.  In a future
commit, we'll instead use nodemon to restart the server when files
change.

You may notice that this commit allows us to move some inline imports to
the top of their respective files.  Top-level imports had to be avoided
for modules with side effects, because clearRequireCache could render
those stale:

 * Suppose ModuleABC is imported in the parent/master/primary server
   process.
 * A worker is forked from the parent process.  Like the parent, it has
   ModuleABC in scope.
 * clearRequireCache is invoked in the worker.
 * A component imports ModuleABC.  Since require.cache was cleared, it's
   re-imported.
 * The component now references a different copy of ModuleABC than the
   worker process.

With clearRequireCache removed, this is no longer any concern.