Sucrase ES module loader
It'd be nice to have a way to quickly execute scripts in Node.js that
may have Flow or JSX syntax without having to build them with Webpack
first (which requires a whole 'nother configuration).

@babel/node is the obvious way, but it doesn't support executing ES
modules.  To do that, we have to write an ES module loader for Node.js,
which is currently behind an experimental flag (but isn't totally
inappropiate for development).

Given that we have to write our own loader, I went with Sucrase instead
of Babel to perform the transformation, as it's much faster and more
suited for this purpose.

The code is based on the coffeescript-loader example by Geoffrey Booth
and contributors:
https://github.com/nodejs/loaders-test/tree/main/coffeescript-loader

We probably won't use this for the template renderer -- at least outside
of development -- because it makes more sense to build that with Webpack
instead (something I will do in a subsequent commit).  But this can be
useful for t/selenium, webpack configuration (imagine having to write a
webpack config for our webpack config), and any other Node.js scripts we
have that we don't want a build step for.