Build root/server.js with Webpack
Prior to this commit, we were restricted on how we could write
root/server.js and any files it imports:

 * No Flow syntax (other than comment syntax)
 * No JSX
 * No ES modules
 * No ProvidePlugin imports

Because root/server.js was executed directly in Node.js, all files had
to conform with normal JavaScript rules and syntax.  (This is why
certain files have had comments saying not to use these features.)

This commit builds root/server.js with Webpack, which means our server
code can now be written without restrictions.  The issue with this is
how to reload the server when things change.

 * In development, we want the server to reload whenever changes to
   components or the server itself are made.  For this, we now use
   nodemon to watch root/static/build/server.js for changes and restart
   the server in its entirety.

 * In production, we want the server to reload *gracefully* when DBDefs
   changes and SIGHUP is received.  For this, we use the start_server
   superdaemon from Server::Starter, which is what we already use for
   plackup in production.  It works perfectly fine for Node.js too by
   specifying the file descriptor to use in an environment variable.

I still think it's useful to be able to run root/server.js directly
without a build step (at least for development), but this is true of
other modules too.  While this commit may seem like a step backwards in
that respect, it's not an unachievable goal:

 * Recent versions of Node.js can execute ES modules directly without
   any experimental flag.

 * ./bin/sucrase-node can be used to execute ES modules that use Flow or
   JSX syntax.

As a last step, I think we should consider removing the ProvidePlugin,
because it ties our code too closely to Webpack.  It does make using
certain utilities more convenient, but Flow's IDE integration (related
to autoimports) has also recently improved to where this isn't as much
of a time-saver as before.