MBS-12497: Drop AC redirects when last use is removed
We automatically remove empty ACs after the last usage
is removed, using the dec_ref_count function.
The artist_credit_gid_redirect table has FKs to artist_credit though,
and it does not affect this process in any way.
That means that when an AC with only 1 usage has a redirect,
any edit that would remove it is blocked because
the redirect is not being removed and the associated FK still exists.

I talked with mwiencek and we decided to implement an intermediate
table where rows can be entered when the ref_count gets to 0,
rather than immediately deleting them. We then wait 7 days to see
if the artist credit got reused (ref_count went above 0 again)
in the meantime. This avoids losing redirects from an AC
that would get used again soon after, for example because it is being
removed from one recording and then added to another.

After 7 days the artist credit gets removed, including any redirects.
If it did get reused in the meantime, we just drop it from the
temporary table, obviously without removing the AC.

The temporary table setup should also work with the alternative
tracklist code in the future, which is why the table is generic
and stores a table name rather than being specifically AC-related.