MBS-12403 (III): Add primary entity type diagrams
These diagrams are defined for all the primary entity types of the
2022 Q2 database schema, that is database schema sequence 27.

For each of these diagrams, the main table is highlighted, other primary
entity tables are shortened (except for `release_group` in release
diagram because there is a lot of overlaps with the `release` table),
materialized tables (if any) are included too and marked as such.

The graphs and images are generated using the command `make`.