MBS-12520: Don't fail on removed annotations
When implementing this I forgot that if an entity is removed,
so are all annotations for it, so the old annotation call on the edit
will return undef.

I considered giving a different "This annotation was removed"
message in this case, but honestly it does not seem useful
for the editor since there's no diff to check anyway.