MBS-11146: Reviews tab for ratings + CB reviews (#1732)
As shown on TV, er... talked about during Summit '20.

This creates a tab for ratings, which we didn't previously have
but makes sense in any case IMO. The tab is named Reviews
and also displays the CB reviews.

We don't really want to display reviews without a body
(rating only reviews) so I request ones with text explicitly.
We do claim there's no reviews when there's just rating-only ones,
but I feel that's fair when we're talking about 'writing a review'.
We could add a third call for ratings but I feel this isn't really
useful enough to justify three calls. Ideally eventually CB will
support getting the average rating for an entity and its recent
and top reviews in one call.

Flow is unhappy because it doesn't realize
entityProperties.ratings limits entities to RatableT and so on,
so I added two FlowIgnores for that.