Dockerfile.website: only build client JS
In production, we can't build the server JS until the container starts
and DBDefs.pm is copied in.  This is because DBDefs.js/mjs is no longer
a Webpack external since aaedd692424e14051c1a96431e6272a8e8c01e69.  It's
built directly into the server bundle.

While it's technically possible to keep DBDefs.mjs an external, with it
being an ES module, we'd also have to output the server bundle as ES
modules.  I believe this is still experimental in Webpack.