MBS-12525: Add fallback sucrase loader for Node.js < 16.12.0
The bin/sucrase-node script added in 6aefc88 doesn't work in versions of
Node.js prior to 16.12.0.  (16.13.0 was the first LTS release, and from
the changelog appears to be identical to 16.12.0.)  This is because the
ESM loader hooks API changed.

Since our INSTALL guide doesn't specify that you need an LTS release of
v16, we should support older versions too.  A fallback loader using the
pre-v16.12 API has been added for this cause.