Fix eslint rule import/extensions
Most of these were autofixed with a temporary use of
node/file-extension-in-import as suggested in
https://github.com/import-js/eslint-plugin-import/issues/1749
A few cases (such as the imports that needed /index.js added)
were amended by hand afterwards.

I changed the rule from 'always' to 'ignorePackages'
since otherwise it wanted react-dom/server and
react-dom/client to get .js appended to them.
My understanding is that is not a problem for ESM support.