Add entities.mjs & generate entities.json from it
Importing entities.json into a JS file using Flow produces an any-typed
value.  Defining a Flow type for this JSON and keeping it up-to-date is
much harder than just defining the ENTITIES structure in JS and letting
Flow create a literal object type for it.  We can then generate
entities.json (which is still used by the Perl) from this file using an
added script.