MBS-11303: Access $c via React context APIs
When I updated root/utility/hydrate.js in
9ef8b79b8e5e4f3c41fb8087b744f8ed90f3ff2d, I forgot to also update its
global type signature in root/vars.js.  When I got around to doing that,
it caused a ton of errors.

Trying to express a proper type that accounts for the component
receiving an unsanitized `$c` on the server and a sanitized one on the
client is pretty hard.  I'd rather components only access `$c` via the
React context APIs, which allows us to keep the `hydrate` type simple.
Server components always being passed `$c` through props wasn't very
obvious or documented, anyway.