MBS-11758: Use current_language instead of lang cookie in Node.js
The Node template renderer currently sets its locale from the lang
cookie.  This may conflict with the locale set by Perl, which would not
be based on the lang cookie if it's not set, but rather the
Accept-Language header.  We should respect the locale determined and set
by Perl, which it stores in current_language in the stash.

This fixes an issue where certain strings (anything rendered by Node)
would be displayed in English even if the browser locale was not
English.