MBS-8876 (I): Show the translation hint translated
Both the UI language change notice and the translation hint were shown
in the language from before the UI language change. It makes sense for
the former, but not for the latter.

This patch splits the flash message to show the notice in the previously
used language and the translation hint in the newly selected language.