MBS-12548: Handle conflicting tag names in MBS-12508 script
Prior to this commit, the script for MBS-12508 didn't take into account
that a tag's ID may change if the last usage is removed and then
re-added.  To handle this, the `ON CONFLICT` specifier is changed to
`(name)` rather than `(id)`, and if there's a conflict we take the ID
associated with that name in the temporary table (which is assumed to be
up-to-date from replication).  To avoid any ID conflicts as before, a
separate `DELETE` is performed first.