Autocomplete2: Use item name for input value
Prior to this commit, if an item was selected, we'd fill in the input
with the name of the entity associated with the item.  In fact, the item
name wasn't really used at all except for filtering.  This isn't
desirable in cases where the item name differs from the entity name.
For example, instruments and link attribute types will have their item
names translated, but the entity names will be in English.