MBS-12576: Improve erroring with wrong IDs on CDTOC pages
This fixes two ISEs:
1) passing a too large integer was accepted by looks_like_number,
but caused an ISE because it was an invalid Postgres integer.
This changes to use is_database_row_id, which is
what we meant with the check.
2) passing a theoretically valid ID that is not associated with a medium
caused an ISE when trying to check for $medium->may_have_discids.
This just bad-requests if the medium ID does not exist,
like we already do with cdtoc IDs for sub move.