MBS-12590: Convert cdstub/logged_in to React
This also makes it so that you can reuse the already provided TOC
while adding the release, which seems like an obvious thing
to support. Sadly, since that's a POST, I don't think it's possible
to do so without breaking the translation.
I'm just using StatusPage since this has the same generic
"this is an error" concept as other CD stub errors.
Also, "Add CD Stub" doesn't make sense as a title
for a page which tells you you cannot do that, so I improved that.