MBS-12205: Add "play on ListenBrainz" button
This mostly copies the functionality
of the existing script
(https://gist.github.com/MonkeyDo/6b919302e97dc979f8277b9c71ff82ab).
It shows a "Play on ListenBrainz" button on the right side of releases,
recordings, recording collections and release groups that sends
the user to play the release or a custom grouping of recordings
with BrainzPlayer.
It only shows on overview pages for recording collections and RGs,
but that's also true of the script so hopefully not a problem.
Ideally, LB would support passing an RG or collection MBID directly.

Keep in mind this doesn't hit LB and doesn't know whether LB can
actually play the music (I'm not sure even hitting LB would help
since LB might also not know until it tries).