MBS-10861: Standardize time before auto-deletion to 48 hours
This sets the RemoveEmpty time to 48 hours / 2 days, up from 1 day,
giving users a bit more time to react to edits that would lead
to autoremoval. It also will make it less common to add artists
for a release, get sidetracked and come back the next day to find
the artists are no longer there.

It also sets the RemoveUnreferencedRows time to 48 hours / 2 days,
down from 7 days, since 7 days means artists in any empty artist credits
remain around for over a week before they finally get autoremoved.

48 hours was chosen as this is the minimum time an open
merge/remove edit needs to apply, so manual and automatic
removals will be consistent. An extra benefit to having the same
duration for merges vs removals is that it reduces the temptation
to let a duplicate be autoremoved rather than merging it to ensure
it disappears faster.