MBS-10499: Alert for entity in danger of removal once add edit passes
Sometimes users (especially new ones) add an entity as a votable edit
(by checking the relevant checkbox) and are then not notified
that the entity is in risk of removal, since their add edit
means there's a pending edit keeping the entity around.
Unless they have added more data, though, the entity will be removed
soon after the one pending edit passes.

This checks if the creation edit is the only thing keeping the entity
from being otherwise eligible for cleanup. If that's the case,
it then notifies the user with the same banner anyway.

I'm also adding a test to see if the banners show as expected
for artist at least, since AFAICT we were not testing this at all.

Notice that is_empty no longer checks edits directly.
The RemoveEmpty autoremoval script actively checks
for edits anyway, so there should be no risk this change would
cause the wrong things to be removed.