Simplify relationship editor's work-recording association
This removes `workRecordings` from the state, which associated each work
with a list of recordings it's linked to.  I believe that simply
traversing the recording-work relationships to find these isn't any
slower, and is less complex to maintain.

One thing I wasn't sure about was the `existingWorkRecordings`
assignment in `updateReleaseRelationships`.  If a recording is unlinked
from a work (say, by removing a pending relationship), we want to make
sure that both the new recording's and the unlinked recording's state is
updated.  But I don't think we need `workRecordings` to get the old,
unlinked recordings: they should already be added to `updatedRecordings`
via a REMOVE_RELATIONSHIP update.