MBS-12628: Hide phrases for unselectable reltypes in inline search
No description means a relationship type is a grouping one
that cannot be selected and only helps organize the tree.
As such, it's useless to display the link phrases for these,
since they're not meant to be selected. The phrases are also usually
effectively placeholders since they never get used anyway.