MBS-12550: Don't enter useless annotation edit in merge (#2640)
The entity merge system was trying to merge the annotations
in entities, but if all entities had the same annotation already,
it still tried to enter an edit, triggering the NoChanges exception
added by MBS-12556.

There seems to be no reason why we would even try to enter an edit
in this case, since the annotation won't change. I'd also argue
it's misleading to claim it's the "Result of $type merge" since it's
just the same as before. This just checks the current annotation
in the target entity, and skips entering an edit if the annotation
resulting from the merge process would have the same text.