Make the relationship credit controls less conspicuous
Merge the credited-as input into the previous row so that it rests below
the entity, taking up less space due to there being less
padding/border-spacing from the removed row.  Remove the left-hand label
text to reduce visual noise, and set the input placeholder to "Credited
as" to clarify its purpose.