MBS-12573: Drop unused degraded column from cdtoc
This was apparently used ages ago to store
"a previous algorithm (less accurate) or things imported from freedb
that didn't have the accuracy of our ids", said Rob,
but "we may have dropped all those back in the day
and never nuked the column".
It contains always FALSE, at the moment,
and we seem to use it exactly nowhere.