MBS-11063: Convert event edit form to React
Since event forms do not display the "Ended" section
of the DateRangeFieldset (we do store ended, but it's
meaningless and we don't use it) this adds a way to hide it on request.

We also need to pass time as a child to DateRangeFieldset,
since we want it to be part of the same fieldset as it is now.

This adds a basic React Bubble component. For now, it keeps
using JQuery for positioning, for simplicity - eventually,
we will probably want to drop that too.