Ignore untyped-import of ko for ko.unwrap
There's a bunch of files a lot of our code depends on which can be made
flow strict (or at least strict-local) except that they use knockout
to unwrap observables from the release editor.
This tells Flow to ignore that specific import (which we know
currently works) so we can get the rest of the benefits of
the stronger typing in these files.