Refactor: Fix Entity role’s name for pending edits
`Entity::Role::Editable` was unclear because it wrongly suggested that
entity can be edited through its own /edit page, or has an _edit table,
or is allowed to be edited by the current user, whereas it just brings
pending edits which can be made to a connected editable entity instead.
Thus renaming as `Entity::Role::PendingEdits`.