MBS-12664: Convert the Register page to React
I added a 'hidden' class rather than setting style because
I was seeing issues with hydrating the styles, but actually
it does seem a lot cleaner to do this anyway.
We might want to use that elsewhere where we pass
a style explicitly now.

I'm not debouncing the email username checker because
I don't think that's particularly useful and just complicates
the code IMO.