MBS-12668: Add URL cleanup for Bugs!
Bugs! is a Korean music streaming platform that also
allows purchasing albums for download.
The URL cleanup rules remove unnecessary segments and
convert mobile URLs to their desktop counterparts.

Since it may happen that albums are only for streaming, we keep the option
to set just the streamingpaid relationship to be on the safe side.