Require `_hash_to_row` for `Data::Role::GID`
`_hash_to_row` used to be implicitly required because it was only
implemented by the descendants of `CoreEntity` package. Now that it has
been replaced by roles, it can be explicitly required.

Track had no `_hash_to_row` because it overrides `_insert_hook_make_row`
which is calling it. As it is now required and since it has an
equivalent `_create_row`, this method is renamed as `_hash_to_row`.

There was only one other equivalent method named `_create_row` left in
Data packages, it was in Medium and is renamed too for consistency.