Actually send users to the right page on linktype delete
This was clearly intended to send users back to the same
type combination page, but since stash->{types} was not loaded here
(this doesn't use type_specific) we were hitting warnings
and ending up at the main rels page instead.
This actually fixes the code to do what was intended,
solving the warnings while at it.