Autocomplete2: Change useLayoutEffect to useEffect
useLayoutEffect produces a warning if you try to run in on the server:

"Warning: useLayoutEffect does nothing on the server, because its effect
cannot be encoded into the server renderer's output format. This will
lead to a mismatch between the initial, non-hydrated UI and the intended
UI. To avoid this, useLayoutEffect should only be used in components
that render exclusively on the client. See
https://reactjs.org/link/uselayouteffect-ssr for common fixes."

This warning is harmless in our case, because the effect in question
doesn't affect the initial rendered output, but is still annoying.

Replacing it with useEffect silences the warning and seems to work fine;
the scroll position of the menu is still updated without any flickering
that I can see.