MBS-12685: Avoild ISE because of concurrent rating transactions
With the previous implementation, we first checked inside a transaction
whether a rating existed, then we tried to insert it if not. While that
seemed sensible, it was causing issues when concurrent
transactions happened - both would check that the rating
did not yet exist, then one would insert it, and the second
would ISE because it now existed.

To avoid this (and limit the number of queries while at it)
this does an INSERT ON CONFLICT DO UPDATE
instead - we always want to either INSERT or UPDATE anyway,
so we don't need to check anything in advance.
This should ensure that if two requests are sent in quick succession,
the first inserts the rating and the second seamlessly updates it.

Added a basic test for the /rate page while at it, since we didn't
seem to have a test that checked whether the ratings were applied,
just that they were allowed/disallowed as needed.