Follow Perl::Critic::Policy::Variables::ProhibitConditionalDeclarations
https://metacpan.org/pod/Perl::Critic::Policy::Variables::ProhibitConditionalDeclarations

We had a bunch of these, but mwiencek said that he tries
to avoid the pattern anyway because of the same reason
the rule suggests to: it's not obvious what it does when
the condition is not true.
In most cases I just changed this to an assignment inside an if.