SEC-900: Upgrade nodemon to use fixed minimatch (#2716)
For website rendering nodemon@2.0.19 was used, which depends on
vulnerable minimatch@3.0.4 (CVE-2022-3517). Even though it seems
unlikely that it might affect the MusicBrainz website, upgrading to
nodemon@2.0.20 makes it use the fixed minimatch@3.1.2 instead.

Note: Other dependencies to minimatch@3.0.4 are used in a development
context only (even glob@7.0.0 via @babel/cli@7.14.8 and shelljs@0.8.5).