MBS-10614: Remove edit note / vote restrictions for beginners
As agreed on the 2022-11-07 team meeting.
These restrictions were introduced before we had a proper
reporting / moderation system, and hopefully they are not needed
anymore. They also make life a lot harder as a good beginner editor.

The restrictions that still allow for beginner editors after this
are the following:
a) They are not allowed to disable banner alerts when they receive
new edit notes on their edits. This seems sensible, since we want
to train beginners to be responsive to edit notes.
b) Their bio and URL are hidden from logged-out users
to limit spam effectiveness.

In addition, edits can still be filtered by whether the editor
is a beginner and the editor report for admins still works.

While doing this I also simplified the description for beginners
in generateUserTypesList - the voting thing is no longer true
so it seems fine to just reuse the string we have elsewhere.