Consistently use named constants for HTTP responses
Following https://metacpan.org/pod/Perl::Critic::Policy::ValuesAndExpressions::ProhibitMagicNumbers
is a lot of work and there's plenty of cases where
it might not even make a lot of sense (such as with IDs in tests).
That said, giving the named values for HTTP constants
rather than the number (something we already did in many files)
is certainly a lot more clear, so I tried to at least
make that consistent.