Follow Perl::Critic::Policy::Modules::ProhibitConditionalUseStatements
https://metacpan.org/pod/Perl::Critic::Policy::Modules::ProhibitConditionalUseStatements

This is apparently just misleading - the package is loaded anyway
when marked use. require should mean it's only loaded if actually needed.
That said, it seems there's no particular reason why the two
conditional use statements we have need to be conditional so this
just moves them to the top of the file with all others.