SEC-123: Upgrade @babel packages to drop kind-of
For website rendering the @babel packages that were used indirectly
depended on vulnerable kind-of@6.0.2 (CVE-2022-37601). Even though it is
unknown how it might affect the MusicBrainz website, upgrading to the
latest @babel packages as below makes these not to use kind-of anymore.

    yarn upgrade --scope @babel --latest --exact

Note that @babel/register@7.18.9 still depends on clone-deep@4.0.1 which
in turn depends on kind-of@6.0.2 or above so kind-of@6.0.3 is actually used.

Note finally that the vulnerable kind-of@3.2.2 still in-use is required
by cliui only which isn’t used for website rendering in production.