Validate invalid session ID cookies earlier
If a request contains an invalid (as in malformed) session ID cookie
paired with a `beta=on` cookie, the request will die before the beta
redirect response is returned and result in a closed connection on the
gateway.

In daf0647c3fdeebd1cc857c14678a619ebd04a3cf we started catching "invalid
session ID" errors in `Controller::Root::begin` and logging them to
Sentry.  The `beta=on` issue arises because a redirect is issued
immediately instead of dispatching the request through
`Controller::Root`.

The potential for an invalid ID to be returned is now handled at the
source by wrapping `get_session_id`.  Since this is called very early in
the dispatch process, we can't render an HTML bad-request page because
there's no view available, so a text/plain error is returned instead.
This is fine because 100% of requests with malformed session IDs are
spam.

I no longer think we should log these errors to Sentry for the same
reason, and because having them in the Docker logs is enough for
forensics.